requirements.txt
requests
pytest
pytest-check
pymysql
pytest-html
appium-python-client
pytest-selenium
